Fix for distributed=True and ngpu > 1
